Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-alignment-style-changes-005-expected.txt

CONSOLE MESSAGE: line 190: No valid data-* attributes found in selector list : .after
É
É
É

PASS .before 1 
PASS .before 2 
PASS .before 3 
